chore(labels): move funcs to approriate place
